+++ routes/validators/treatmentValidators.ts
@@ -4,7 +4,7 @@ import validationErrorCheck from "./validationErrorCheck.js"
 const treatmentPostValidator = [
   body('start_date').notEmpty().isISO8601().toDate(),
   body('title').notEmpty(),
-  body('patient_id').notEmpty().isInt(),
+  body('patient_id').notEmpty().isInt().toInt(),
   validationErrorCheck
 ]
